// src/app/models/clinica.model.ts
export class ClinicaModel {
    uid?: string;
    id?: string;
    clave: string;
    nombre: string;
    razon_social: string;
    numero_de_consultorios: number;
    numero_de_quirofanos: number;
    telefono_fijo: string;
    email: string;
    calle: string;
    numero_exterior: string;
    numero_interior: string;
    colonia: string;
    delegacion_o_municipio: string;
    codigo_postal: string;
    ciudad: string;
    estado: string;
    mapTitle: string;
    mapComment: string;
    mapLat: number;
    mapLng: number;
}